// src/store/enums/store-messages.enum.ts
export enum StoreSuccessMessages {
  CREATED = "انبار ایجاد شد",
  FOUND = "انبار پیدا شد",
  LISTED = "لیست انبار ها دریافت شد",
  UPDATED = "انبار با موفقیت بروز رسانی شد",
  DELETED = "انبار با موفقیت حذف شد",
}

export enum StoreErrorMessages {
  DUPLICATE = "انبار تکراری است",
  NOT_FOUND = "انبار پیدا نشد",
}